just a couple changes to button behaviour on the workspace dialog, changed slot from released() to clicked(), not very important
There was a problem with not being able to create a new workspace if you chose a file
with the file browser in the WorkspaceDialog, a Q_ASSERT in the method getType() in the
XMLElementInfo class would fail when the ServerList object was initialized in the Workspace constructor.  
I believe the problem is because ServerList stores a QList of type XMLElemnt*.  
To fix this issue I made two changes:

1) registered the class XMLElement in the registerClasses() method in Workspace
2) called registerClasses() in the createWorkspace() method so that all of the required
classes are registered before the Workspace constructor is called

These two changes seem to have fixed the issue because creating a new workspace now works
correctly, and you are able to add servers with the ServerWizard that will persist through
restarts of the gui.

+++ muse_gui/infra/ssh/SshChannel.cpp
@@ -39,9 +39,11 @@
 #include "SshChannel.h"
 #include <QFile>
 #include <QFileInfo>
-#include <stdio.h>
 #include <QMessageBox>
 
+#include <cstdio>
+
+#include <fstream>
 #include <array>
 #include <algorithm>
 
@@ -65,11 +67,11 @@ SshChannel::copy(const QString &srcDir, const QString &destDirectory,
     size_t readToBuffer;
     QString remoteFilePath = destDirectory +
             (destDirectory.endsWith("/") ? destFileName : "/" + destFileName);
-    unsigned long notRead = fileInfo.size();
+    size_t notRead = fileInfo.size();
 
     // Send a file via scp. The mode parameter must only have permissions!
     channel = libssh2_scp_send(session, remoteFilePath.toStdString().c_str(), mode,
-                              (unsigned long)notRead);
+                               notRead);
 
     // If the channel isn't open.
     if (!channel) {
@@ -80,13 +82,13 @@ SshChannel::copy(const QString &srcDir, const QString &destDirectory,
 
     char buffer[1024];
     char* ptr;
-    size_t bytesWritten;
+    ssize_t bytesWritten;
 
     while (notRead > 0) {
         readToBuffer = fread(buffer, 1, sizeof(buffer), srcFile);
         ptr = buffer;
 
-        while (readToBuffer != 0){
+        while (readToBuffer != 0) {
             //write the same data over and over, until error or completion
             bytesWritten = libssh2_channel_write(channel, ptr, readToBuffer);
